+++ package.json
@@ -64,5 +64,7 @@
   "files": [
     "dist"
   ],
-  "dependencies": {}
+  "dependencies": {
+    "node-sass": "^4.14.1"
+  }
 }
